+++ ext/spl/spl_heap.c
@@ -535,7 +535,7 @@ static inline HashTable* spl_heap_object_get_debug_info(zend_class_entry *ce, ze
 }
 /* }}} */
 
-static HashTable *spl_heap_object_get_gc(zend_object *obj, zval **gc_data, int *gc_data_count) /* {{{ */
+static HashTable *spl_heap_object_get_gc(zend_object *obj, zval **gc_data, ptrdiff_t *gc_data_count) /* {{{ */
 {
 	spl_heap_object *intern = spl_heap_from_obj(obj);
 	*gc_data = (zval *) intern->heap->elements;
@@ -545,7 +545,7 @@ static HashTable *spl_heap_object_get_gc(zend_object *obj, zval **gc_data, int *
 }
 /* }}} */
 
-static HashTable *spl_pqueue_object_get_gc(zend_object *obj, zval **gc_data, int *gc_data_count) /* {{{ */
+static HashTable *spl_pqueue_object_get_gc(zend_object *obj, zval **gc_data, ptrdiff_t *gc_data_count) /* {{{ */
 {
 	spl_heap_object *intern = spl_heap_from_obj(obj);
 	*gc_data = (zval *) intern->heap->elements;
