+++ ext/xml/xml.c
@@ -154,7 +154,7 @@ static PHP_GINIT_FUNCTION(xml);
 
 static zend_object *xml_parser_create_object(zend_class_entry *class_type);
 static void xml_parser_free_obj(zend_object *object);
-static HashTable *xml_parser_get_gc(zend_object *object, zval **table, int *n);
+static HashTable *xml_parser_get_gc(zend_object *object, zval **table, ptrdiff_t *n);
 static zend_function *xml_parser_get_constructor(zend_object *object);
 
 static zend_string *xml_utf8_decode(const XML_Char *, size_t, const XML_Char *);
@@ -414,7 +414,7 @@ static void xml_parser_free_obj(zend_object *object)
 	zend_object_std_dtor(&parser->std);
 }
 
-static HashTable *xml_parser_get_gc(zend_object *object, zval **table, int *n)
+static HashTable *xml_parser_get_gc(zend_object *object, zval **table, ptrdiff_t *n)
 {
 	xml_parser *parser = xml_parser_from_obj(object);
 	*table = &parser->object;
